# laravel_bajas/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bajas laborales",
        "description": "Bajas",
        "version": "12.17.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/index": {
            "get": {
                "tags": [
                    "Bajas"
                ],
                "summary": "Lista de bajas activas y todas las bajas",
                "operationId": "42246c76c1a7eaac5f63bc7a28672d6b",
                "responses": {
                    "200": {
                        "description": "tabla",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "datosBajasActivas": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Baja"
                                            }
                                        },
                                        "datosBajasTodas": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Baja"
                                            }
                                        },
                                        "datosTrabajadores": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Trabajador"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/baja/{id_baja}": {
            "get": {
                "tags": [
                    "BajasEspecíficas"
                ],
                "summary": "Obtener una baja por su ID",
                "description": "Mostrar los detalles de una baja",
                "operationId": "c84bb2d8cd0ce1c44944fed9d9da7b51",
                "parameters": [
                    {
                        "name": "id_baja",
                        "in": "path",
                        "description": "ID de la baja",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la baja específica",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Baja"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Baja no encontrada"
                    }
                }
            }
        },
        "/insertarBaja": {
            "post": {
                "tags": [
                    "insertarbaja"
                ],
                "summary": "Crear una nueva baja laboral",
                "description": "Crear una nueva baja",
                "operationId": "a15f1cb3cf11c1ec83ff04b6bc158ad7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dni",
                                    "fecha_inicio",
                                    "fecha_fin",
                                    "motivo"
                                ],
                                "properties": {
                                    "dni": {
                                        "description": "DNI del trabajador",
                                        "type": "string",
                                        "example": "11111111A"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-01"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-01"
                                    },
                                    "motivo": {
                                        "type": "string",
                                        "example": "Accidente laboral"
                                    },
                                    "comentario": {
                                        "type": "string",
                                        "example": "Rotura del ligamento cruzado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Baja creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Baja"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en los datos enviados"
                    },
                    "409": {
                        "description": "Baja duplicada"
                    },
                    "500": {
                        "description": "Error en servidor"
                    }
                }
            }
        },
        "/cancelarBaja": {
            "get": {
                "tags": [
                    "cancelarBaja"
                ],
                "summary": "cancelo una baja",
                "description": "Cancelo una baja",
                "operationId": "df2fb62d2adf5011409fdd9e0f2b5bf2",
                "parameters": [
                    {
                        "name": "id_baja",
                        "in": "path",
                        "description": "id de la baja a cancelar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Baja cancelada y correo enviado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "baja": {
                                            "properties": {
                                                "fecha_inicio": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-06-06"
                                                },
                                                "fecha_fin": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-06-10"
                                                },
                                                "motivo": {
                                                    "type": "string",
                                                    "example": "Prueba manual"
                                                },
                                                "dni": {
                                                    "type": "string",
                                                    "example": "12345678A"
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Ana"
                                                },
                                                "apellido1": {
                                                    "type": "string",
                                                    "example": "H"
                                                },
                                                "apellido2": {
                                                    "type": "string",
                                                    "example": "H"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Baja no encontrada"
                    },
                    "500": {
                        "description": "Error al cancelar la baja o enviar el correo"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Baja": {
                "title": "Baja",
                "required": [
                    "id_baja",
                    "fecha_inicio",
                    "fecha_fin",
                    "duracion",
                    "motivo",
                    "trabajador"
                ],
                "properties": {
                    "id_baja": {
                        "type": "integer",
                        "example": 1
                    },
                    "fecha_inicio": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-06"
                    },
                    "fecha_fin": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-07-06"
                    },
                    "duracion": {
                        "type": "string",
                        "example": "1 mes"
                    },
                    "motivo": {
                        "type": "string",
                        "example": "Accidente laboral"
                    },
                    "comentario": {
                        "type": "string",
                        "example": "Caída"
                    },
                    "trabajador": {
                        "type": "integer",
                        "example": 1
                    },
                    "estado": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "Notificacion": {
                "title": "Notificación",
                "required": [
                    "baja",
                    "fecha"
                ],
                "properties": {
                    "baja": {
                        "type": "integer",
                        "example": 1
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-09"
                    }
                },
                "type": "object"
            },
            "Trabajador": {
                "title": "Trabajador",
                "required": [
                    "id_trabajador",
                    "dni",
                    "nombre",
                    "apellido1",
                    "apellido2",
                    "correo"
                ],
                "properties": {
                    "id_trabajador": {
                        "type": "integer",
                        "example": 1
                    },
                    "dni": {
                        "type": "string",
                        "example": "11111111A"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Aaaa"
                    },
                    "apellido1": {
                        "type": "string",
                        "example": "Apellido1Aaa"
                    },
                    "apellido2": {
                        "type": "string",
                        "example": "Apellido2Aaaa"
                    },
                    "ano_nac": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "aaaaa@gmail.com"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "C/ AAAA"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "111111111"
                    },
                    "departamento": {
                        "type": "string",
                        "example": "IT"
                    },
                    "puesto": {
                        "type": "string",
                        "example": "Desarrollador"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Bajas",
            "description": "Bajas"
        },
        {
            "name": "BajasEspecíficas",
            "description": "BajasEspecíficas"
        },
        {
            "name": "insertarbaja",
            "description": "insertarbaja"
        },
        {
            "name": "cancelarBaja",
            "description": "cancelarBaja"
        }
    ]
}